Only saving image

+++ generateTicket.php
@@ -18,19 +18,22 @@ $ticketImage = imagecreatefrompng('resources/ticket.png');
 imagealphablending($ticketImage, false);
 imagesavealpha($ticketImage, true);
 
-imagecopymerge($ticketImage, $ticketCode, 0, 0, 0, 0, 250, 40, 100);
+$done = imagecopymerge($ticketImage, $ticketCode, 0, 0, 0, 0, 250, 40, 100);
 
+if ($done) {
+    $saveto = "generatedTickets/" . $idTicket . ".png";
+    if(file_exists($saveto)){
+        unlink($saveto);
+    }
+    $fp = fopen($saveto,'x');
+    fwrite($fp, $ticketImage);
+    fclose($fp);
+}
 //header('Content-Type: image/png');
 //imagepng($ticketImage);
 
 
-$saveto = "generatedTickets/" . $idTicket . ".png";
-if(file_exists($saveto)){
-    unlink($saveto);
-}
-$fp = fopen($saveto,'x');
-fwrite($fp, $ticketCode);
-fclose($fp);
+
 
 function grab_image($url, $idTicket){
     $ch = curl_init ($url);
